+++ src/article_extractor/models.py
