+++ tests/services/article_classification/test_corruption_adapter_integration.py
@@ -6,12 +6,12 @@ Run sparingly to avoid API costs.
 import pytest
 from datetime import datetime, timezone
 
-from src.services.article_classification.models import (
+from src.article_classification.models import (
     ClassificationInput,
     ClassificationResult,
     ClassifierType,
 )
-from src.services.article_classification.agents.corruption_classifier import (
+from src.article_classification.agents.corruption_classifier import (
     CorruptionClassifier,
 )
